Implemented ValidateForm component and imported to all user profile changes Issue #(sprint 2)
